created new ontext class
