+++ Vireo_VS/VireoCommandLine.vcxproj.filters
@@ -148,6 +148,9 @@
     <ClCompile Include="..\source\core\CloseReference.cpp">
       <Filter>VireoSource\Core</Filter>
     </ClCompile>
+    <ClCompile Include="..\source\core\DebuggingContext.cpp">
+      <Filter>VireoSource\Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\source\include\ConversionTable.def">
@@ -281,6 +284,9 @@
       <Filter>VireoSource\Include</Filter>
     </ClInclude>
     <ClInclude Include="..\source\include\DebuggingToggles.h">
+      <Filter>VireoSource\Include</Filter>
+    </ClInclude>
+    <ClInclude Include="..\source\include\DebuggingContext.h">
       <Filter>VireoSource\Include</Filter>
     </ClInclude>
   </ItemGroup>
