+++ make-it/Makefile
@@ -15,7 +15,7 @@ OUTPUT_EXE=$(OUTPUT_DIR)/esh
 OUTPUT_TEST_EXE=$(OUTPUT_DIR)/esh-test
 
 COMMANDLINE = main.cpp
-CORE = Array.cpp Assert.cpp CEntryPoints.cpp CloseReference.cpp ControlRef.cpp Date.cpp DualTypeEqual.cpp DualTypeOperation.cpp DualTypeConversion.cpp DualTypeVisitor.cpp EventLog.cpp Events.cpp ExecutionContext.cpp GenericFunctions.cpp JavaScriptStaticRef.cpp JavaScriptDynamicRef.cpp MatchPat.cpp Math.cpp NumericString.cpp Platform.cpp Queue.cpp RefNum.cpp String.cpp StringUtilities.cpp Synchronization.cpp TDCodecLVFlat.cpp TDCodecVia.cpp Thread.cpp TimeFunctions.cpp Timestamp.cpp TypeAndDataManager.cpp TypeAndDataReflection.cpp TypeDefiner.cpp TypeTemplates.cpp UnitTest.cpp  Variants.cpp VirtualInstrument.cpp Waveform.cpp
+CORE = Array.cpp Assert.cpp CEntryPoints.cpp CloseReference.cpp ControlRef.cpp DebuggingContext.cpp Date.cpp DualTypeEqual.cpp DualTypeOperation.cpp DualTypeConversion.cpp DualTypeVisitor.cpp EventLog.cpp Events.cpp ExecutionContext.cpp GenericFunctions.cpp JavaScriptStaticRef.cpp JavaScriptDynamicRef.cpp MatchPat.cpp Math.cpp NumericString.cpp Platform.cpp Queue.cpp RefNum.cpp String.cpp StringUtilities.cpp Synchronization.cpp TDCodecLVFlat.cpp TDCodecVia.cpp Thread.cpp TimeFunctions.cpp Timestamp.cpp TypeAndDataManager.cpp TypeAndDataReflection.cpp TypeDefiner.cpp TypeTemplates.cpp UnitTest.cpp  Variants.cpp VirtualInstrument.cpp Waveform.cpp
 UNITTEST = RefNumTest.cpp
 IO = FileIO.cpp DebugGPIO.cpp HttpClient.cpp JavaScriptInvoke.cpp
