+++ source/core/DebuggingContext.cpp
@@ -1,0 +1,14 @@
+// Copyright(c) 2020 National Instruments
+// SPDX-License-Identifier: MIT
+
+/*! \file
+	\brief helper foe debugging class
+ */
+#include "debuggingContext.h"
+
+#if kVireoOS_emscripten
+#include <emscripten.h>
+#endif
+
+namespace Vireo {
+}  // namespace Vireo
